requirements.txt
Django==2.2.7
django-crispy-forms==1.8.1
djangorestframework==3.10.3
pytz==2019.3
sqlparse==0.3.0
